Refactoring the app
This commit refactors the structure of the app, bringing it closer to
the implementation as seen in the GoLang Otel collector.

To facilitate this I've decoupled components of a pipeline (receivers,
processors and exporters) by using unbounded mpsc channels for comms
between them.

I've also added thiserror and anyhow, but I suspect I can remove anyhow
now and just progress with thiserror.